<refactor> 파티 컴포넌트 디렉터리 구조 변경, 참조 경로 수정
