+++ src/components/Party/PartyBoard.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Typography } from '@mui/material';
-import { BOARD_PARTY } from '../constants/Party';
+import { BOARD_PARTY } from '../../constants/Party';
 import PartyFilter from './PartyFilter';
 import PartyList from './PartyList';
