+++ src/components/Party/PartyMemberList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { List, Divider } from '@mui/material';
 import PartyMemberListItem from './PartyMemberListItem';
-import { PartyMemberListProps } from '../types/Party';
+import { PartyMemberListProps } from '../../types/Party';
 
 function PartyMemberList({ members }: PartyMemberListProps) {
 	return (
