+++ src/components/Party/PartyMemberListItem.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Stack, Avatar, Typography, Chip, Button, Box } from '@mui/material';
-import { PartyMember } from '../types/Party';
-import { PARTY_LIST_ITEM } from '../constants/Party';
+import { PartyMember } from '../../types/Party';
+import { PARTY_LIST_ITEM } from '../../constants/Party';
 
 function PartyMemberListItem({ member }: { member: PartyMember }) {
 	return (
